Move html adapters to different package
