+++ rosa-iiif/rosa-iiif-presentation-core/src/main/java/rosa/iiif/presentation/core/html/GraphHtmlAdapter.java
@@ -1,4 +1,4 @@
-package rosa.iiif.presentation.core.util;
+package rosa.iiif.presentation.core.html;
 
 import com.google.inject.Inject;
 import rosa.archive.model.Book;
